+++ packages/larva-patterns/modules/table/table.prototype.js
@@ -1,0 +1,7 @@
+module.exports = {
+	table_background_color_class: '',
+	table_classes: 'lrv-u-overflow-x-auto',
+	table_color_class: '',
+	table_markup: 'Table content.',
+	table_text_align_class: '',
+};